Removed close() from ui namespace and added sample
Partners were being confused by the fact that close() method doesn't exist in the ui namespace. This may be implemented in the future, but currently doesn't exist, so fixed this. Also added and e2e sample to show the key elements that need to be in place to make a basic scenario work.