Document maximum custom property value length
Custom properties are truncated to 255 16-bit UTF-16 code-units in the desktop edition of Word, while this limitation does not appear to exist in Word Online. This truncation appears to be unicode-unaware: if the 255th character and 256th characters form a surrogate pair, the truncation will truncate so that the value ends with a high surrogate, without the matching low surrogate. This commit documents this surprising behavior.